Condenser Improvement + Added Graph Import
Performance update for wikipedia xml condenser.  No longer rewrites the
variables when checking for extra links in the line. Instead it starts
looking later in the variable.

Created code that can import the condensed wikipedia stuff into the
snap.py graphs. Uses a 'for elements in graph' search on each to term to
determine if  each term is unique before adding it to the graph. I
suspect there is a more efficient way to do this. Run time on creating
this graph looks like it is in the hundreds of hours. Currently I am
running it. Let me know if you create something faster.