It Lives!!!
So the trimmed xml file we made was all sorts of fucked, I updated the
code to work with it properly a number of issues have been fixed:

1) Article titles were imported with \r\n at their end, removed this and
the names now match link names.

2) If there was a : or | anywhere on the page the link text to be
assigned was the start of the current link all the way up to the : or |.
This has been fixed, probably not in the most efficient way.

3) Added automatic dead end deletion.  If a new to the hash table
article has no children it's value in the hash table is replaced with
the next title. There are a number of wiki internal articles or articles
that only link to wiki talk that this is effected by this change. This
does require at minimum one variable reset and one if check per cycle.
This only catches articles that have just been added to the hash table
so this may not be worth it as we can pull the dead end articles out in
post.

Updated code comments come Monday after this is done running.